backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pillow==10.4.0
tensorflow==2.17.0
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.5.2
python-jose==3.3.0
passlib==1.7.4
aiofiles==24.1.0
pydantic==2.9.0
matplotlib==3.9.2
seaborn==0.13.2
kaggle==1.6.17